refactor(order): 修改订单状态类型为int
- 将订单状态字段类型从Byte改为Integer
- 更新相关接口参数类型为int
- 修改状态常量赋值方式，去除byte强制转换
- 调整服务层方法参数类型以保持一致性
- 更新数据库实体类中状态字段的类型定义

+++ src/main/java/com/studio/settlement/bean/po/ActualOrdersPo.java
@@ -93,11 +93,11 @@ public class ActualOrdersPo implements Serializable {
 
     @Schema(description = "订单状态（1:待服务 2:服务中 3:已完成 4:已取消）")
     @TableField("order_status")
-    private Byte orderStatus;
+    private Integer orderStatus;
 
     @Schema(description = "结算状态（0:未结算 1:已结算）")
     @TableField("settlement_status")
-    private Byte settlementStatus;
+    private Integer settlementStatus;
 
     @Schema(description = "服务开始时间")
     @TableField("service_start_time")
